tinflate: Catch LZ offsets pointing before the start of output buffer.
We do this only in "memory output buffer" mode, and it takes to use
destStart field (which was present in TINF_DATA, but wasn't really
used).

In streaming mode, with a dictionary buffer, we don't add overhead
of performing analog of this check: checking that dictionary offset
accesses actually filled in part of the dictionary buffer. In this
mode, such offsets can't lead to accessing mmeory outside the
dictionary buffer, they only lead to exposing older contents of
this buffer, so it's up to the caller to initialize it before calling
uzlib.