tgunzip: Don't allow to decompress more data than specified in gzip trailer.
As we allocate output buffer based on the size specified in the trailer,
we must take precautions to not overrun this buffer in case the actual
compressed stream contains more data. We also need to detect such a
mismatch and report it as an error. (Error reporting will be a bit
funky, "Error during decompression: 0", where 0 is TINF_OK, meaning
that there're more data in the stream, whereas the expected code
is TINF_DONE.)